+++ src/compiler/compile/nodes/shared/Context.ts
@@ -1,5 +1,5 @@
 import { x } from 'code-red';
-import { Node, Identifier, Expression } from 'estree';
+import { Node, Identifier, Expression, PrivateIdentifier } from 'estree';
 import { walk } from 'estree-walker';
 import is_reference, { NodeWithPropertyDefinition } from 'is-reference';
 import { clone } from '../../../utils/clone';
@@ -7,7 +7,16 @@ import Component from '../../Component';
 import flatten_reference from '../../utils/flatten_reference';
 import TemplateScope from './TemplateScope';
 
-export interface Context {
+export type Context = DestructuredVariable | ComputedProperty;
+
+interface ComputedProperty {
+	type: 'ComputedProperty';
+  property_name: Identifier;
+  key: Expression | PrivateIdentifier;
+}
+ 
+interface DestructuredVariable {
+	type: 'DestructuredVariable'
 	key: Identifier;
 	name?: string;
 	modifier: (node: Node) => Node;
@@ -25,8 +34,8 @@ export function unpack_destructuring({
 }: {
 	contexts: Context[];
 	node: Node;
-	modifier?: Context['modifier'];
-	default_modifier?: Context['default_modifier'];
+	modifier?: DestructuredVariable['modifier'];
+	default_modifier?: DestructuredVariable['default_modifier'];
 	scope: TemplateScope;
 	component: Component;
 	context_rest_properties: Map<string, Node>;
@@ -35,12 +44,14 @@ export function unpack_destructuring({
 
 	if (node.type === 'Identifier') {
 		contexts.push({
+			type: 'DestructuredVariable',
 			key: node as Identifier,
 			modifier,
 			default_modifier
 		});
 	} else if (node.type === 'RestElement') {
 		contexts.push({
+			type: 'DestructuredVariable',
 			key: node.argument as Identifier,
 			modifier,
 			default_modifier
@@ -112,22 +123,32 @@ export function unpack_destructuring({
 				const key = property.key;
 				const value = property.value;
 
-				let property_name: any;
 				let new_modifier: (node: Node) => Node; 
 
 				if (property.computed) {
-					// TODO: If the property is computed, ie, { [computed_key]: prop }, the computed_key can be any type of expression.
+					// e.g { [computedProperty]: ... }
+					const property_name = component.get_unique_name('computed_property');
+
+					contexts.push({
+						type: 'ComputedProperty',
+						property_name,
+						key
+					});
+
+					new_modifier = (node) => x`${modifier(node)}[${property_name}]`;
+					used_properties.push(x`${property_name}`);
 				} else if (key.type === 'Identifier') {
 					// e.g. { someProperty: ... }
-				  property_name = key.name;
+					const property_name = key.name;
 					new_modifier = (node) => x`${modifier(node)}.${property_name}`;
+					used_properties.push(x`"${property_name}"`);
 				} else if (key.type === 'Literal') {
 					// e.g. { "property-in-quotes": ... } or { 14: ... }
-					property_name = key.value;
+					const property_name = key.value;
 					new_modifier = (node) => x`${modifier(node)}["${property_name}"]`;
+					used_properties.push(x`"${property_name}"`);
 				}
 				
-			  used_properties.push(x`"${property_name}"`);
 				if (value.type === 'AssignmentPattern') {
 					// e.g. { property = default } or { property: newName = default }
 					const n = contexts.length;
@@ -174,7 +195,9 @@ function update_reference(
 ): Node {
 	const find_from_context = (node: Identifier) => {
 		for (let i = n; i < contexts.length; i++) {
-			const { key } = contexts[i];
+			const cur_context = contexts[i];
+			if (cur_context.type !== 'DestructuredVariable') continue; 
+			const { key } = cur_context;
 			if (node.name === key.name) {
 				throw new Error(`Cannot access '${node.name}' before initialization`);
 			}