+++ src/compiler/compile/render_dom/invalidate.ts
@@ -19,7 +19,6 @@ export function invalidate(renderer: Renderer, scope: Scope, node: Node, names: 
 				!variable.hoistable &&
 				!variable.global &&
 				!variable.module &&
-				!variable.is_reactive_static &&
 				(
 					variable.referenced ||
 					variable.subscribable ||