+++ src/compiler/compile/utils/a11y.ts
@@ -19,7 +19,8 @@ const non_interactive_roles = new Set(
 				// 'toolbar' does not descend from widget, but it does support
 				// aria-activedescendant, thus in practice we treat it as a widget.
 				// focusable tabpanel elements are recommended if any panels in a set contain content where the first element in the panel is not focusable.
-				!['toolbar', 'tabpanel'].includes(name) &&
+				// 'generic' is meant to have no semantic meaning.
+				!['toolbar', 'tabpanel', 'generic'].includes(name) &&
 				!role.superClass.some((classes) => classes.includes('widget'))
 			);
 		})
@@ -31,7 +32,11 @@ const non_interactive_roles = new Set(
 );
 
 const interactive_roles = new Set(
-	non_abstract_roles.filter((name) => !non_interactive_roles.has(name))
+	non_abstract_roles.filter((name) => 
+		!non_interactive_roles.has(name) &&
+		// 'generic' is meant to have no semantic meaning.
+		name !== 'generic'
+	)
 );
 
 export function is_non_interactive_roles(role: ARIARoleDefinitionKey) {
@@ -66,6 +71,24 @@ export function is_hidden_from_screen_reader(tag_name: string, attribute_map: Ma
 	if (!aria_hidden.is_static) return true;
 	const aria_hidden_value = aria_hidden.get_static_value();
 	return aria_hidden_value === true || aria_hidden_value === 'true';
+}
+
+export function has_disabled_attribute(attribute_map: Map<string, Attribute>) {
+	const disabled_attr = attribute_map.get('disabled');
+	const disabled_attr_value = disabled_attr && disabled_attr.get_static_value();
+	if (disabled_attr_value) {
+		return true;
+	}
+
+	const aria_disabled_attr = attribute_map.get('aria-disabled');
+	if (aria_disabled_attr) {
+		const aria_disabled_attr_value = aria_disabled_attr.get_static_value();
+		if (aria_disabled_attr_value === true) {
+			return true;
+		}
+	}
+
+	return false;
 }
 
 const non_interactive_element_role_schemas: ARIARoleRelationConcept[] = [];