+++ test/runtime/samples/attribute-escape/main.svelte
@@ -1,0 +1,1 @@
+<textarea value={`test'"></textarea><script>alert('BIM');</script>`} />